Localise and decrease openfermion-dependency in qchem tests (#2593)
