+++ tests/qchem/test_hamiltonians.py
