+++ tests/qchem/test_observable_hf.py
