+++ tests/qchem/test_spin.py
